Forward whole recieptblobs slice to ccv data
