+++ protocol/common/storageaccess/inmemory_storage_test.go
@@ -55,13 +55,16 @@ func TestInMemoryOffchainStorage_StoreCCVData(t *testing.T) {
 			SourceVerifierAddress: verifierAddress,
 			DestVerifierAddress:   []byte("0x4567"),
 			CCVData:               []byte("signature1"),
+			BlobData:              []byte("blob1"),
 			Timestamp:             time.Now().Unix(),
-			ReceiptBlob: common.ReceiptWithBlob{
-				Issuer:            verifierAddress,
-				DestGasLimit:      200000,
-				DestBytesOverhead: 50,
-				Blob:              []byte("blob1"),
-				ExtraArgs:         []byte{},
+			ReceiptBlobs: []common.ReceiptWithBlob{
+				{
+					Issuer:            verifierAddress,
+					DestGasLimit:      200000,
+					DestBytesOverhead: 50,
+					Blob:              []byte("blob1"),
+					ExtraArgs:         []byte{},
+				},
 			},
 			Message: createTestMessage(100, 1, 2),
 		},
@@ -73,13 +76,16 @@ func TestInMemoryOffchainStorage_StoreCCVData(t *testing.T) {
 			SourceVerifierAddress: verifierAddress,
 			DestVerifierAddress:   []byte("0x4567"),
 			CCVData:               []byte("signature2"),
+			BlobData:              []byte("blob2"),
 			Timestamp:             time.Now().Unix() + 1,
-			ReceiptBlob: common.ReceiptWithBlob{
-				Issuer:            verifierAddress,
-				DestGasLimit:      300000,
-				DestBytesOverhead: 75,
-				Blob:              []byte("blob2"),
-				ExtraArgs:         []byte("test"),
+			ReceiptBlobs: []common.ReceiptWithBlob{
+				{
+					Issuer:            verifierAddress,
+					DestGasLimit:      300000,
+					DestBytesOverhead: 75,
+					Blob:              []byte("blob2"),
+					ExtraArgs:         []byte("test"),
+				},
 			},
 			Message: createTestMessage(101, 1, 2),
 		},
@@ -276,13 +282,16 @@ func TestInMemoryOffchainStorage_GetCCVDataByMessageID(t *testing.T) {
 			SourceVerifierAddress: verifierAddress,
 			DestVerifierAddress:   []byte("0x4567"),
 			CCVData:               []byte("signature"),
+			BlobData:              []byte("blob"),
 			Timestamp:             time.Now().Unix(),
-			ReceiptBlob: common.ReceiptWithBlob{
-				Issuer:            verifierAddress,
-				DestGasLimit:      150000,
-				DestBytesOverhead: 30,
-				Blob:              []byte("blob"),
-				ExtraArgs:         []byte{},
+			ReceiptBlobs: []common.ReceiptWithBlob{
+				{
+					Issuer:            verifierAddress,
+					DestGasLimit:      150000,
+					DestBytesOverhead: 30,
+					Blob:              []byte("blob"),
+					ExtraArgs:         []byte{},
+				},
 			},
 			Message: createTestMessage(100, 1, 2),
 		},
@@ -482,12 +491,15 @@ func TestInMemoryOffchainStorage_ReaderWriterViews(t *testing.T) {
 			SourceVerifierAddress: verifierAddress,
 			DestVerifierAddress:   []byte("0x4567"),
 			CCVData:               []byte("signature1"),
-			ReceiptBlob: common.ReceiptWithBlob{
-				Issuer:            verifierAddress,
-				DestGasLimit:      250000,
-				DestBytesOverhead: 60,
-				Blob:              []byte("blob1"),
-				ExtraArgs:         []byte("extra"),
+			BlobData:              []byte("blob1"),
+			ReceiptBlobs: []common.ReceiptWithBlob{
+				{
+					Issuer:            verifierAddress,
+					DestGasLimit:      250000,
+					DestBytesOverhead: 60,
+					Blob:              []byte("blob1"),
+					ExtraArgs:         []byte("extra"),
+				},
 			},
 			Message: createTestMessage(100, 1, 2),
 		},
