+++ verifier/commit/validation.go
@@ -3,11 +3,9 @@ package commit
 import (
 	"bytes"
 	"fmt"
-	"time"
 
 	"github.com/smartcontractkit/chainlink-ccv/protocol/common"
 	"github.com/smartcontractkit/chainlink-ccv/verifier/types"
-	"github.com/smartcontractkit/chainlink-ccv/verifier/utils"
 )
 
 // ValidateMessage validates a verification task message using the new format
@@ -43,37 +41,4 @@ func ValidateMessage(verificationTask *types.VerificationTask, verifierOnRampAdd
 	}
 
 	return nil
-}
-
-// CreateCCVData creates CCVData from verification task, signature, and blob using the new format
-func CreateCCVData(verificationTask *types.VerificationTask, signature []byte, verifierBlob []byte, sourceVerifierAddress common.UnknownAddress) (*common.CCVData, error) {
-	message := verificationTask.Message
-	messageID, err := message.MessageID()
-	if err != nil {
-		return nil, fmt.Errorf("failed to compute message ID: %w", err)
-	}
-	// Find the receipt blob that corresponds to our source verifier address
-	verifierIndex, err := utils.FindVerifierIndexBySourceAddress(verificationTask, sourceVerifierAddress)
-	if err != nil {
-		return nil, fmt.Errorf("failed to find verifier index: %w", err)
-	}
-
-	// Get the receipt blob at the verifier index
-	if verifierIndex >= len(verificationTask.ReceiptBlobs) {
-		return nil, fmt.Errorf("verifier index %d is out of bounds for receipt blobs (length: %d)", verifierIndex, len(verificationTask.ReceiptBlobs))
-	}
-	receiptBlob := verificationTask.ReceiptBlobs[verifierIndex]
-
-	return &common.CCVData{
-		MessageID:             messageID,
-		SequenceNumber:        message.SequenceNumber,
-		SourceChainSelector:   message.SourceChainSelector,
-		DestChainSelector:     message.DestChainSelector,
-		SourceVerifierAddress: sourceVerifierAddress,
-		DestVerifierAddress:   common.UnknownAddress{}, // Will be set by the caller if needed
-		CCVData:               signature,
-		Timestamp:             time.Now().UnixMicro(), // Unix timestamp in microseconds
-		Message:               message,
-		ReceiptBlob:           receiptBlob,
-	}, nil
 }
